Fix bad config key names

+++ config/services.php
@@ -45,7 +45,7 @@ return [
 	'x' => [
 		'access_token' => env('X_ACCESS_TOKEN'),
 		'access_token_secret' => env('X_ACCESS_TOKEN_SECRET'),
-		'account_id' => env('X_ACCOUNT_ID'),
+		'account_id' => (int)env('X_ACCOUNT_ID'),
 		'api_key' => env('X_API_KEY'),
 		'api_secret' => env('X_API_SECRET'),
 		'bearer_token' => env('X_BEARER_TOKEN'),
